// web/src/api/models/TypesMacroInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesMacroInfo
 */
export interface TypesMacroInfo {
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    applied_at?: string;
    /**
     * 
     * @type {number}
     * @memberof TypesMacroInfo
     */
    apply_count: number;
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    created_at: string;
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    created_by: string;
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TypesMacroInfo
     */
    name: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof TypesMacroInfo
     */
    sections?: Array<string>;
}

/**
 * Check if a given object implements the TypesMacroInfo interface.
 */
export function instanceOfTypesMacroInfo(value: object): value is TypesMacroInfo {
    if (!('apply_count' in value) || value['apply_count'] === undefined) return false;
    if (!('created_at' in value) || value['created_at'] === undefined) return false;
    if (!('created_by' in value) || value['created_by'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function TypesMacroInfoFromJSON(json: any): TypesMacroInfo {
    return TypesMacroInfoFromJSONTyped(json, false);
}

export function TypesMacroInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesMacroInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'applied_at': json['applied_at'] == null ? undefined : json['applied_at'],
        'apply_count': json['apply_count'],
        'created_at': json['created_at'],
        'created_by': json['created_by'],
        'description': json['description'],
        'id': json['id'],
        'name': json['name'],
        'sections': json['sections'] == null ? undefined : json['sections'],
    };
}

export function TypesMacroInfoToJSON(json: any): TypesMacroInfo {
    return TypesMacroInfoToJSONTyped(json, false);
}

export function TypesMacroInfoToJSONTyped(value?: TypesMacroInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'applied_at': value['applied_at'],
        'apply_count': value['apply_count'],
        'created_at': value['created_at'],
        'created_by': value['created_by'],
        'description': value['description'],
        'id': value['id'],
        'name': value['name'],
        'sections': value['sections'],
    };
}